refactor: 클래스 네이밍 변경, 패키지 이동
