+++ src/main/java/calculator/Application.java
@@ -1,12 +1,12 @@
 package calculator;
 
-import calculator.controller.StringCalculator;
+import calculator.controller.CustomCalculator;
 import calculator.view.InputView;
 import calculator.view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
-        StringCalculator calculator = new StringCalculator(new InputView(), new OutputView());
+        CustomCalculator calculator = new CustomCalculator(new InputView(), new OutputView());
         calculator.run();
     }
 }
